+++ app/components/annotations/annotations-directives.js
@@ -129,11 +129,24 @@ angular.module('annotations')
        * on error.
        * @param {object} annotation - The annotation to be deleted.
        */
-      scope.deleteAnnotation = function (annotation) {
+      var deleteAnnotation = function (annotation) {
         AnnotationsService.deleteAnnotation(
           annotation,
           deleteAnnotationSuccess.bind(undefined, annotation),
           deleteAnnotationError);
+      };
+
+      /**
+       * Delete a notification after request for confirmation.
+       * @param {object} annotation - The annotation to be deleted.
+       */
+      scope.deleteAnnotation = function (annotation) {
+        notie.confirm(
+            gettextCatalog.getString(
+              "Are you sure you want to delete this annotation?"),
+            gettextCatalog.getString("Yes"),
+            gettextCatalog.getString("No"),
+            deleteAnnotation.bind(undefined, annotation));
       };
     };
 
@@ -214,9 +227,9 @@ angular.module('annotations')
 angular.module('annotations')
   .directive('annotationsMake',
              ['AnnotationsService', '$window', 'gettextCatalog', 'notie',
-              'user',
+              'user', 'State',
               function (AnnotationsService, $window, gettextCatalog, notie,
-                        user) {
+                        user, State) {
 
     var link = function (scope, element, attrs) {
 
@@ -254,6 +267,7 @@ angular.module('annotations')
        */
       var createAnnotationSuccess = function(value, responseHeaders){
         scope.annotations.splice(0, 0, value);
+        scope.resetForm();
         AnnotationsService.refreshAnnotationLayer();
       };